api/requirements.txt
fastapi
uvicorn[standard]
numpy
scipy
supabase_py
